commit 22-12

+++ front-end/src/components/Payment/Payment.js
@@ -3,20 +3,35 @@ import "./payment.css";
 import axios from "../../config/axios";
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate } from "react-router-dom";
+import { EditOutlined, DeleteOutlined, HomeOutlined } from "@ant-design/icons";
+import { Button, Modal } from "antd";
+import { toast } from "react-toastify";
 function Payment() {
   const navigate = useNavigate();
   const selector = useSelector((state) => state);
   const CartID = selector.CartID;
-  const userId = JSON.parse(localStorage.getItem("info")).id;
+  const userId = JSON.parse(localStorage.getItem("info"))._id;
+  const addressLocal = JSON.parse(localStorage.getItem("info")).address;
+  const [address, setAddress] = useState(addressLocal[0]);
   const [myCart, setCart] = useState([]);
-
+  const [isModalOpen, setIsModalOpen] = useState(false);
+  const showModal = () => {
+    setIsModalOpen(true);
+  };
+  const handleOk = () => {
+    setIsModalOpen(false);
+  };
+  const handleCancel = () => {
+    setIsModalOpen(false);
+  };
   const getCart = async () => {
-    let info = localStorage.getItem("info");
+    let info = JSON.parse(localStorage.getItem("info"));
     if (info) {
-      const resp = await axios.get(`/api/cart/mycart/${JSON.parse(info).id}`);
+      const resp = await axios.get(`/api/cart/mycart/${info._id}`);
       setCart(resp.data.listProduct);
     }
   };
+
   let sum = 0;
   myCart.forEach((element) => {
     sum += element.quantity * element.productId.price;
@@ -26,18 +41,22 @@ function Payment() {
     getCart();
   }, []);
   const handlePayment = async () => {
-    const resp = await axios.post(`/api/order/neworder`, {
-      userId: userId,
-      listProducts: myCart,
-      total: sum,
-      address: "123",
-    });
-    if (resp.status === 200) {
-      let clearCart = await axios.delete(`/api/cart/clear/${CartID}`);
-      console.log(clearCart);
+    if (!address) {
+      toast.warning("Thiếu thông tin địa chỉ");
+    } else {
+      const resp = await axios.post(`/api/order/neworder`, {
+        userId: userId,
+        listProducts: myCart,
+        total: sum,
+        address: address,
+      });
+
+      if (resp.status === 200) {
+        let clearCart = await axios.delete(`/api/cart/clear/${CartID}`);
+        toast.success("Vui lòng chờ xác nhận đơn hàng");
+        navigate("/profile/order/waiting");
+      }
     }
-    alert("Vui lòng chờ xác nhận đơn hàng");
-    navigate("/profile/order");
   };
   return (
     <div className="payment_content">
@@ -71,9 +90,10 @@ function Payment() {
         <div className="box-border">
           <div className="payment_delivery">
             <span>Giao hàng đến</span>
-            <div className="add">
-              113 Chiến Thắng, P. Văn Quán, Hà Đông, Hà Nội, Vietnam
-            </div>
+            <p>
+              {address}
+              <EditOutlined onClick={showModal} />
+            </p>
           </div>
           <div className="note-time">
             Thời gian tiếp nhận đơn hàng trực tuyến từ <br />
@@ -117,6 +137,30 @@ function Payment() {
           </div>
         </div>
       </div>
+      <Modal
+        title="ĐỊA ĐIỂM ĐÃ LƯU"
+        open={isModalOpen}
+        onOk={handleOk}
+        onCancel={handleCancel}
+        width={1000}
+        bodyStyle={{ minHeight: 250 }}
+        footer={null}
+      >
+        <div className="address-content">
+          {addressLocal?.map((item, index) => (
+            <div
+              className="address-content-item"
+              key={index}
+              onClick={() => {
+                handleCancel();
+                setAddress(item);
+              }}
+            >
+              <span>{item}</span>
+            </div>
+          ))}
+        </div>
+      </Modal>
     </div>
   );
 }
